refactor: move SpinalSimBmbLengthSpliterTester
